Move package.json to root of repo
There are two conflicting issues that necessitate this:

1. We want to allow projects to use this instead of the file:// syntax to make it
   easier to handle upgrades (removing the need to run `yarn upgrade jest-preset-shadow`).
   We also want to keep this repo private until we open source.
2. When adding a git URL, you cannot specify a subdirectory.

Additionally, the recommended solution of using GitPkg (https://gitpkg.vercel.app) doesn't
work because the repo is private.

To work around this, until we publish the package on NPM, we can just keep `jest-preset-shadow`'s
`package.json` file at the repo root. Later we will move it back to the package directory.