Simplify matcher setup and fix bug related to negated matchers
By moving more of the logic into the compile stage, it's possible to
simplify a lot of the internals related to matchers. Specifically,
instead of setting the dynamic `*inside-is?` and `*is-body-negated?`
bindings inside of `cljest.core/is` and reading these values inside
of `cljest.matchers/make-matcher` and `make-optional/matcher`, which
created an implicit contract between the two, we have a `defmatcher`
macro which generates a function that has explicit `:jest-matcher`
metadata which is analyzed at compile time. This lets all of the
`expect` code exist within `is`, making it easier to reason about
how things work at the expense of more things moving to compile time.

Additionally, the previous implementation did not correctly support
negated matchers, and was only discovered after some manual testing
and exploration. This change fixes this bug.

- Support alises when running cljest

Previously, if you had an alias such as `test` which you were starting
`clj` with, while the code would correctly compile, but Jest wouldn't
find any extra paths that alias added.