Properly handle `let` bindings.
Previously, if you had a let binding that depended on a previous binding,
it wouldn't work because the `async` macro would expand into a `js/Promise.all`
call with the values, and the symbols as the bindings for the next `.then` call.
This meant that if one of the values in `Promise.all` depended on a previous
binding symbol, it wouldn't be able to resolve it. This commit fixes this by
generating a new `let` call for each pair if any of the binding pairs has
`await`.

Fixes #30.